Moved tower functionality from main
